requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-wtf==1.1.1
wtforms==3.0.1
mysql-connector-python==8.1.0
werkzeug==2.3.7
tensorflow==2.20.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
bcrypt==4.0.1
python-dotenv==1.0.0
